backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
python-dotenv
numpy==2.3.2
pandas==2.3.1
scikit-learn==1.7.1
scipy==1.16.1
tensorflow==2.17.0
joblib==1.5.1
threadpoolctl==3.6.0
yfinance==0.2.65
requests==2.32.4
beautifulsoup4==4.13.4
textblob
plotly==6.2.0
openpyxl
frozendict==2.4.6
typing_extensions==4.14.1
packaging==25.0
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
pytz==2025.2
tzdata==2025.2
multitasking==0.0.12
narwhals==2.0.1
peewee==3.18.2
platformdirs==4.3.8
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
redis>=5.0.0
websockets==15.0.1
soupsieve==2.7
curl_cffi==0.12.0
protobuf==6.31.1
pycparser==2.22
six==1.17.0